requirements.txt
elpy==1.9.0
flake8==2.4.1
hypothesis==1.11.4
importmagic==0.1.3
jedi==0.9.0
mccabe==0.3.1
mock==1.3.0
pbr==1.8.1
peewee==2.6.4
pep8==1.5.7
polib==1.0.7
py==1.4.30
pyflakes==0.8.1
pytest==2.8.2
PyYAML==3.11
requests==2.8.1
six==1.10.0
vcrpy==1.7.4
wrapt==1.10.5
